requirements.txt
click==8.1.3
colorama==0.4.6
colorgram.py==1.2.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
Werkzeug==2.2.2
